forced building type id conversion to int
git-svn-id: https://svn.urbansim.org/src/trunk@17139 1513fa5d-83ce-422e-ae47-6967bb0fb594

+++ urbansim_parcel/development_project_proposal_component/building_sqft_per_job.py
@@ -18,8 +18,8 @@ class building_sqft_per_job(Variable):
     def compute(self,  dataset_pool):
         sqft_per_job = dataset_pool.get_dataset("building_sqft_per_job")
         proposal_components = self.get_dataset()
-        zones = proposal_components.get_attribute("zone_id")
-        type_ids = proposal_components.get_attribute("building_type_id")
+        zones = proposal_components.get_attribute("zone_id").astype('int32')
+        type_ids = proposal_components.get_attribute("building_type_id").astype('int32')
         building_sqft_per_job_table = sqft_per_job.get_building_sqft_as_table(zones.max(), type_ids.max())
         return building_sqft_per_job_table[zones, type_ids]
